==> 0 - preprocessing_tocsv.ipynb <==
import numpy as np
import pandas as pd 
pd.options.mode.chained_assignment = None


# Preprocessing

Due to the very high size of the dataset, we did the preprocessing once here and saved it to csv files

import pandas as pd

traindf = pd.read_csv('data/rotten_tomatoes_movie_reviews.csv')



traindf.head()
y_train = pd.DataFrame(traindf['scoreSentiment'])
x_train = pd.DataFrame()
x_train['text']=traindf['reviewText'].astype(str)
display(x_train.head())
y_train.head()


X= x_train['text']
y=y_train['scoreSentiment']
y.value_counts()

from preprocessing import preprocessed


data = preprocessed(x_train[:35000])


display(data.head())
data.to_csv("data/x_train.csv", index=False)
y_train[:35000].to_csv("data/y_train.csv", index=False)
