==> 1- explortory_data_analysis.ipynb <==
import pandas as pd

# Data Exploration

I encountered an issue with matplotlib so i will only use plotly.express

All outputs are cleared due to file size.

data = pd.read_csv('data/rotten_tomatoes_movie_reviews.csv')
moviesinfo = pd.read_csv('data/rotten_tomatoes_movies.csv')

data.head()

data_all = data.merge(moviesinfo, on='id', how = 'left')

data_all.head()

len(data_all)

import plotly.express as px
class_labels = ["Positive", "Negative"]
category_orders = {"Sentiment": class_labels}

print(data_all["scoreSentiment"].value_counts(normalize=True)*100)

fig = px.histogram(data_all, x="scoreSentiment", title="Sentiment distribution", color = "scoreSentiment", category_orders=category_orders,text_auto=True)
fig.show()

We have twice as many positive values than negative ones, we will maybe have overfitting problems

data_all.columns

data_all['word_count'] = data_all['reviewText'].apply(lambda x: len(str(x).split()))

# Create a histogram
fig = px.histogram(data_all, x='word_count', nbins=30, title='Distribution of Number of Words in ReviewText')

# Customize the layout if needed
fig.update_layout(
    xaxis_title='Number of Words',
    yaxis_title='Frequency',
    showlegend=False  # You can set this to True if you want a legend
)

# Show the plot
fig.show()

We see that we have a lot of empty reviews

fig = px.histogram(
    data_all, 
    x='word_count', 
    color='scoreSentiment', 
    nbins=30, 
    title='Distribution of Number of Words in ReviewText by Sentiment',
    labels={'word_count': 'Number of Words', 'scoreSentiment': 'Sentiment'},
    category_orders={'scoreSentiment': ['positive', 'negative']}
)

# Customize the layout if needed
fig.update_layout(
    yaxis_title='Frequency',
    showlegend=True
)

# Show the plot
fig.show()

The repartition of the values does not seem to depend on the number of words

import plotly.express as px
import pandas as pd


top_movies = data_all['title'].value_counts().nlargest(10).index
top_movies_data = data_all[data_all['title'].isin(top_movies)]
color_palette = {'POSITIVE': 'green', 'NEGATIVE': 'red'}
fig = px.bar(
    top_movies_data,
    x='title',
    color='scoreSentiment',
    title='Top 10 Most Frequent Movies and Sentiment Distribution',
    labels={'title': 'Movie Title', 'scoreSentiment': 'Sentiment'},
    category_orders={'scoreSentiment': ['POSITIVE', 'NEGATIVE']},
    color_discrete_map=color_palette
)

fig.update_layout(
    xaxis_title='Movie Title',
    yaxis_title='Count',
    showlegend=True
)
fig.show()


import plotly.express as px
import pandas as pd


top_movies = data_all['director'].value_counts().nlargest(5).index
top_movies_data = data_all[data_all['director'].isin(top_movies)]
color_palette = {'POSITIVE': 'green', 'NEGATIVE': 'red'}
fig = px.bar(
    top_movies_data,
    x='director',
    color='scoreSentiment',
    title='Top 5 Most Frequent directors and Sentiment Distribution',
    labels={'title': 'Directore Name', 'scoreSentiment': 'Sentiment'},
    category_orders={'scoreSentiment': ['POSITIVE', 'NEGATIVE']},
    color_discrete_map=color_palette
)

fig.update_layout(
    xaxis_title='Movie Title',
    yaxis_title='Count',
    showlegend=True
)
fig.show()

Here we have the most present directors and movies and their ratio of positive/negative reviews

A better analysis could be interesting but would not really change the approach for the NLP classification
